manage listings

+++ routes/web.php
@@ -24,39 +24,38 @@ Route::get('/', [ListingController::class, 'index']);
 Route::get('/listings/create', [ListingController::class, 'create'])->middleware('auth');
 
 // store listing request
-Route::post('/listings', 
-[ListingController::class, 'store'])->middleware('auth');
+Route::post('/listings', [ListingController::class, 'store'])->middleware('auth');
 
 // Show edit Form
 Route::get('/listings/{listing}/edit', [ListingController::class, 'edit']);
 
 // Update listing
-Route::put('/listings/{listing}', 
-[ListingController::class, 'update'])->middleware('auth');
+Route::put('/listings/{listing}', [ListingController::class, 'update'])->middleware('auth');
 
 // Delete listing
-Route::delete('/listings/{listing}', 
-[ListingController::class, 'delete'])->middleware('auth');
+Route::delete('/listings/{listing}', [ListingController::class, 'delete'])->middleware('auth');
+
+// Manage Listings
+Route::get('/listings/manage', [ListingController::class, 'manage'])->middleware('auth');
 
 // Single listing
 Route::get('/listings/{listing}', [ListingController::class, 'show']);
 
 // show Register/create Form
-Route::get('/register', 
-[UserController::class, 'register'])->middleware('guest');
+Route::get('/register', [UserController::class, 'register'])->middleware('guest');
 
 // Create new User
 Route::post('/users', [UserController::class, 'store']);
 
 // Log user out
-Route::post('/logout', 
-[UserController::class, 'logout'])->middleware('auth');
+Route::post('/logout', [UserController::class, 'logout'])->middleware('auth');
 
 // Show login form
 Route::get('/login', [UserController::class, 'login'])->name('login')->middleware('guest');
 
 // Log in User
 Route::post('/users/authenticate', [UserController::class, 'authenticate']);
+
 
 // Route::get('/posts/{id}', function($id){
 //     return response('Posts '.$id);
